requirements.txt
pipdeptree

watchdog
streamlit
openai
python-dotenv
